Capitalizing Java File Names
Per java naming best practices
