+++ app/src/main/java/com/example/fond/activities/MainActivity.java
@@ -7,18 +7,14 @@ import androidx.fragment.app.FragmentManager;
 
 import android.os.Bundle;
 import android.view.MenuItem;
-import android.widget.FrameLayout;
-import android.widget.Toast;
-import android.util.Log;
 
-import com.example.fond.fragments.profileFragment;
-import com.example.fond.fragments.savedRecipesFragment;
-import com.example.fond.fragments.searchRecipeFragment;
-import com.example.fond.fragments.userFeedFragment;
+import com.example.fond.fragments.ProfileFragment;
+import com.example.fond.fragments.SavedRecipesFragment;
+import com.example.fond.fragments.SearchRecipeFragment;
+import com.example.fond.fragments.UserFeedFragment;
 import com.google.android.material.bottomnavigation.BottomNavigationView;
 
 import com.example.fond.R;
-import com.parse.ParseUser;
 
 public class MainActivity extends AppCompatActivity {
     public static final String TAG = "MainActivity";
@@ -33,10 +29,10 @@ public class MainActivity extends AppCompatActivity {
         bottomNavigationView = findViewById(R.id.bottom_navigation);
         final FragmentManager fragmentManager = getSupportFragmentManager();
 
-        final Fragment userFeedFragment = new userFeedFragment();
-        final Fragment searchRecipeFragment = new searchRecipeFragment();
-        final Fragment savedRecipesFragment = new savedRecipesFragment();
-        final Fragment profileFragment = new profileFragment();
+        final Fragment userFeedFragment = new UserFeedFragment();
+        final Fragment searchRecipeFragment = new SearchRecipeFragment();
+        final Fragment savedRecipesFragment = new SavedRecipesFragment();
+        final Fragment profileFragment = new ProfileFragment();
 
 
         bottomNavigationView.setOnNavigationItemSelectedListener(
