+++ app/src/main/res/layout/fragment_saved_recipes.xml
@@ -3,7 +3,7 @@
     xmlns:tools="http://schemas.android.com/tools"
     android:layout_width="match_parent"
     android:layout_height="match_parent"
-    tools:context=".fragments.savedRecipesFragment">
+    tools:context=".fragments.SavedRecipesFragment">
 
     <include
         layout="@layout/toolbar_main"/>
