+++ app/src/main/res/layout/fragment_user_feed.xml
@@ -4,7 +4,7 @@
     xmlns:tools="http://schemas.android.com/tools"
     android:layout_width="match_parent"
     android:layout_height="match_parent"
-    tools:context=".fragments.userFeedFragment">
+    tools:context=".fragments.UserFeedFragment">
 
 
     <include
